refactor: split design route resolution helpers and add coverage
