+++ apps/store/src/features/design/components/chat/chat-panel.test.tsx
@@ -9,7 +9,7 @@ import { forwardRef, useImperativeHandle } from "react";
 import userEvent from "@testing-library/user-event";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import { beforeEach, describe, expect, it, vi } from "vitest";
-import { buildAnalysisReuseKey } from "@/entities/design/api/analysis-reuse-key";
+import { buildAnalysisReuseKey } from "@/entities/design";
 import { ChatPanel } from "@/features/design/components/chat/chat-panel";
 import { useDesignChatStore } from "@/features/design/store/design-chat-store";
