+++ apps/store/src/shared/lib/imagekit.ts
@@ -8,7 +8,9 @@ export const IMAGEKIT_URL_ENDPOINT =
 export const IMAGEKIT_PUBLIC_KEY =
   import.meta.env.VITE_IMAGEKIT_PUBLIC_KEY || "";
 
-if (!IMAGEKIT_URL_ENDPOINT) {
+const shouldWarnForMissingEndpoint = import.meta.env.MODE !== "test";
+
+if (!IMAGEKIT_URL_ENDPOINT && shouldWarnForMissingEndpoint) {
   console.warn(
     "ImageKit URL Endpoint가 설정되지 않았습니다. VITE_IMAGEKIT_URL_ENDPOINT 환경 변수를 설정하세요.",
   );
